pyproject.toml
[project]
name = "wiggybot-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pyairtable>=3.3.0",
    "requests>=2.32.5",
]
